+++ chalicelib/processors/CompareCodeProcessor.py
@@ -1,4 +1,4 @@
-from chalicelib.processors.GenericProcessor import GenericProcessor
+from chalicelib.processors.GenericProcessor import GenericProcessor, AnalysisOutputFormat
 from chalicelib.version import API_VERSION
 from chalicelib.usage import OpenAIDefaults
 from chalice import BadRequestError
@@ -12,7 +12,8 @@ class CompareCodeProcessor(GenericProcessor):
             ['system', 'compare-code-original-code-role-system.prompt']],
             None,
             {'model': OpenAIDefaults.boost_default_gpt_model,
-             'temperature': OpenAIDefaults.temperature_terse_and_accurate})
+             'temperature': OpenAIDefaults.temperature_terse_and_accurate},
+            AnalysisOutputFormat.bulletedList, [AnalysisOutputFormat.bulletedList, AnalysisOutputFormat.rankedList, AnalysisOutputFormat.numberedList])
 
     def get_chunkable_input(self) -> str:
         return 'code'