+++ chalicelib/processors/SummaryProcessor.py
@@ -1,6 +1,9 @@
 import math
 
-from chalicelib.processors.GenericProcessor import GenericProcessor, key_ChunkedInputs, key_ChunkPrefix, key_IsChunked, key_NumberOfChunks
+from chalicelib.processors.GenericProcessor import (GenericProcessor,
+                                                    key_ChunkedInputs, key_ChunkPrefix,
+                                                    key_IsChunked, key_NumberOfChunks,
+                                                    AnalysisOutputFormat)
 from chalicelib.version import API_VERSION
 from chalicelib.usage import OpenAIDefaults
 from chalice import BadRequestError
@@ -13,7 +16,8 @@ class SummarizeProcessor(GenericProcessor):
                           ['system', 'summarize-role-system.prompt']],
                          None,
                          {'model': OpenAIDefaults.boost_default_gpt_model,
-                          'temperature': OpenAIDefaults.temperature_medium_with_explanation})
+                          'temperature': OpenAIDefaults.temperature_medium_with_explanation},
+                         AnalysisOutputFormat.prose)
 
     def calculate_input_token_buffer(self, total_max) -> int:
         # we'll leave 90% of the buffer for the input, and 10% for the output